+++ service/directconnect/api.go
@@ -1262,7 +1262,7 @@ func (c *DirectConnect) CreateBGPPeerRequest(input *CreateBGPPeerInput) (req *re
 // you cannot specify custom IPv6 addresses.
 //
 // For a public virtual interface, the Autonomous System Number (ASN) must be
-// private or already whitelisted for the virtual interface.
+// private or already on the allow list for the virtual interface.
 //
 // Returns awserr.Error for service API and SDK errors. Use runtime type assertions
 // with awserr.Error's Code and Message methods to get detailed information about
@@ -3260,13 +3260,21 @@ func (c *DirectConnect) DescribeDirectConnectGatewayAssociationsRequest(input *D
 // DescribeDirectConnectGatewayAssociations API operation for AWS Direct Connect.
 //
 // Lists the associations between your Direct Connect gateways and virtual private
-// gateways. You must specify a Direct Connect gateway, a virtual private gateway,
-// or both. If you specify a Direct Connect gateway, the response contains all
-// virtual private gateways associated with the Direct Connect gateway. If you
-// specify a virtual private gateway, the response contains all Direct Connect
-// gateways associated with the virtual private gateway. If you specify both,
-// the response contains the association between the Direct Connect gateway
-// and the virtual private gateway.
+// gateways and transit gateways. You must specify one of the following:
+//
+//    * A Direct Connect gateway The response contains all virtual private gateways
+//    and transit gateways associated with the Direct Connect gateway.
+//
+//    * A virtual private gateway The response contains the Direct Connect gateway.
+//
+//    * A transit gateway The response contains the Direct Connect gateway.
+//
+//    * A Direct Connect gateway and a virtual private gateway The response
+//    contains the association between the Direct Connect gateway and virtual
+//    private gateway.
+//
+//    * A Direct Connect gateway and a transit gateway The response contains
+//    the association between the Direct Connect gateway and transit gateway.
 //
 // Returns awserr.Error for service API and SDK errors. Use runtime type assertions
 // with awserr.Error's Code and Message methods to get detailed information about
@@ -8223,7 +8231,7 @@ type DescribeDirectConnectGatewayAssociationsInput struct {
 	// The token provided in the previous call to retrieve the next page.
 	NextToken *string `locationName:"nextToken" type:"string"`
 
-	// The ID of the virtual private gateway.
+	// The ID of the virtual private gateway or transit gateway.
 	VirtualGatewayId *string `locationName:"virtualGatewayId" type:"string"`
 }